+++ myBank.h
@@ -6,6 +6,6 @@ void getBalance(int id);//gets id and printd it's balance
 void deposit(int id, double amount);//gets amount to deposit and prints new balance
 void withdraw(int id, double amount);//gets amount to withdraw and prints new balance
 void shut(int id);//closes given account
-void addInterest(double prec);//adds to all accounts interest by given precentage
+void addInterest(int prec);//adds to all accounts interest by given precentage
 void report();//prints all opened accounts balance
 #endif
